Enable starlingx mirror
StarlingX needs to download a variety of rpms and tarballs
from various upstream sources.  Unfortunately the upstream sources
are not always dependable. Either servers go down, are unreachable,
or drop older content that we still depend on.

Our proposed solution is to run our own mirror to capture an
independent copy of the content needed by StarlingX.
For this purpose, a server has been set up at
http://mirror.starlingx.cengn.ca/mirror/centos

The mirror will use deterministic paths derived from the upstream
urls.  Scripts will be used to convert an upstream url to
the mirror's equivalent url (see function url_to_cengn_url in
url_utils.sh)

The mirror will be refreshed daily.  New .lst entries will be
processed at that time.  Processing of changes under yum.repos.d
is not automated by this update.  Expect a follow-up update to
address this issue soon.  These scripts are found under the
'cengn' subdirectory.

Changes are made to the download_mirror.sh script, and it's
supporting scripts.  New arguments have been added to each
script to select the download source.
   -s  StarlingX mirror only
   -S  StarlingX mirror, with upstream source as backup
   -u  Upstream source only
   -U  Upstream source, with StarlingX mirror as backup

You do not need to provide any of these flags.  Continue to
us download_mirror.sh as you always have.  The default
behavior is currently set to '-S', i.e. first try the
StarlingX mirror, with upstream source as backup.
If this proves to place to heavy a load on the existing
server, we might switch the default to '-U', i.e. first
try the upstream source, with StarlingX mirror as backup.
If poor download performance is seen, you might want to try
explicitly adding -U as an argument.

The remaining two options are not recommended for regular use.
Upstream only, i.e. '-u', restores original behaviour, but
you may once again encounter rpms that have aged out, and
been removed from their original repos.  StarlingX only,
i.e. '-s', is vulnerable if a .lst file has been updated,
but the mirror has not yet processed it.

Story: 2003906
Task: 26785

+++ centos-mirror-tools/dl_tarball.sh
@@ -1,5 +1,9 @@
 #!/usr/bin/env bash
 
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
 # The build of StarlingX relies, besides RPM Binaries and Sources, in this
 # repository which is a collection of packages in the form of Tar Compressed
 # files and 3 RPMs obtained from a Tar Compressed file. This script and a text
@@ -13,6 +17,66 @@
 script_path="$(dirname $(readlink -f $0))"
 tarball_file="$script_path/tarball-dl.lst"
 mvn_artf_file="$script_path/mvn-artifacts.lst"
+
+DL_TARBALL_DIR="$(dirname "$(readlink -f "${BASH_SOURCE[0]}" )" )"
+
+source $DL_TARBALL_DIR/url_utils.sh
+
+usage () {
+    echo "$0 [-D <distro>] [-s|-S|-u|-U] [-h] <other_download_list.ini> <save_path> [<force_update>]"
+}
+
+# Permitted values of dl_source
+dl_from_stx_mirror="cengn"
+dl_from_upstream="upstream"
+dl_from_stx_then_upstream="$dl_from_stx_mirror $dl_from_upstream"
+dl_from_upstream_then_stx="$dl_from_upstream $dl_from_stx_mirror"
+
+# Download from what source?
+#   dl_from_stx_mirror = StarlingX mirror only
+#   dl_from_upstream   = Original upstream source only
+#   dl_from_stx_then_upstream = Either source, STX prefered (default)"
+#   dl_from_upstream_then_stx = Either source, UPSTREAM prefered"
+dl_source="$dl_from_stx_then_upstream"
+
+distro="centos"
+
+# Parse out optional arguments
+while getopts "D:hsSuU" o; do
+    case "${o}" in
+        D)
+            distro="${OPTARG}"
+            ;;
+
+        s)
+            # Download from StarlingX mirror only. Do not use upstream sources.
+            dl_source="$dl_from_stx_mirror"
+            ;;
+        S)
+            # Download from StarlingX mirror only. Do not use upstream sources.
+            dl_source="$dl_from_stx_then_upstream"
+            ;;
+        u)
+            # Download from upstream only. Do not use StarlingX mirror.
+            dl_source="$dl_from_upstream"
+            ;;
+        U)
+            # Download from upstream only. Do not use StarlingX mirror.
+            dl_source="$dl_from_upstream_then_stx"
+            ;;
+        h)
+            # Help
+            usage
+            exit 0
+            ;;
+        *)
+            usage
+            exit 1
+            ;;
+    esac
+done
+shift $((OPTIND-1))
+
 
 if [ ! -e $tarball_file -o ! -e $mvn_artf_file ];then
     echo "$download_list does not exist, please have a check!"
@@ -40,19 +104,54 @@ fi
 # Download function using wget command
 
 download_package() {
-    wget --spider $1
-    if [ $? != 0 ]; then
-        echo "$1 is broken"
-    else
-        wget -t 5 --wait=15 $1
+    local upstream_url="$1"
+    local stx_url=""
+    local url=""
+    local rc=1
+
+    stx_url="$(url_to_cengn_url "$upstream_url" "$distro")"
+
+    for dl_src in $dl_source; do
+        case $dl_src in
+            $dl_from_stx_mirror)
+                url="$stx_url"
+                ;;
+            $dl_from_upstream)
+                url="$upstream_url"
+                ;;
+            *)
+                echo "Error: Unknown dl_source '$dl_src'"
+                continue
+                ;;
+        esac
+
+        wget --spider "$url"
         if [ $? != 0 ]; then
-            echo "$1" > "$output_log"
+            echo "Warning: '$url' is broken"
+        else
+            wget -t 5 --wait=15 "$url"
+            if [ $? -eq 0 ]; then
+                rc=0
+                break
+            else
+                echo "Warning: failed to download '$url'"
+                continue
+            fi
         fi
+    done
+
+    if [ $rc != 0 ]; then
+        echo "Error: failed to download '$upstream_url'"
+        echo "$upstream_url" > "$output_log"
     fi
+
+    return $rc
 }
 
 # This script will iterate over the tarball.lst text file and execute specific
 # tasks based on the name of the package:
+
+error_count=0;
 
 for line in $(cat $tarball_file); do
 
@@ -74,6 +173,8 @@ for line in $(cat $tarball_file); do
     tarball_name=$(echo $line | cut -d"#" -f1-1)
     directory_name=$(echo $line | cut -d"#" -f2-2)
     tarball_url=$(echo $line | cut -d"#" -f3-3)
+    method=$(echo $line | cut -d"#" -f4-4)
+    util=$(echo $line | cut -d"#" -f5-5)
 
     # Remove leading '!' if present
     tarball_name="${tarball_name//!/}"
@@ -104,6 +205,12 @@ for line in $(cat $tarball_file); do
         pushd $output_tarball
         if [ "$tarball_name" = "integrity-kmod-e6aef069.tar.gz" ]; then
             download_package $tarball_url
+            if [ $? -ne 0 ]; then
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
             tar xf e6aef069b6e97790cb127d5eeb86ae9ff0b7b0e3.tar.gz
             mv linux-tpmdd-e6aef06/security/integrity/ $directory_name
             tar czvf $tarball_name $directory_name
@@ -111,6 +218,12 @@ for line in $(cat $tarball_file); do
             rm e6aef069b6e97790cb127d5eeb86ae9ff0b7b0e3.tar.gz
         elif [ "$tarball_name" = "mariadb-10.1.28.tar.gz" ]; then
             download_package $tarball_url
+            if [ $? -ne 0 ]; then
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
             mkdir $directory_name
             tar xf $tarball_name --strip-components 1 -C $directory_name
             rm $tarball_name
@@ -122,6 +235,7 @@ for line in $(cat $tarball_file); do
             popd
             tar czvf $tarball_name $directory_name
             rm -rf $directory_name
+            popd    # pushd $directory_name
         # The mvn.repo.tgz tarball will be created downloading a serie of
         # of maven artifacts described in mvn-artifacts file.
         elif [ "$tarball_name" = "mvn.repo.tgz" ]; then
@@ -130,10 +244,48 @@ for line in $(cat $tarball_file); do
                 echo "$mvn_artf_file no found" 1>&2
                 exit 1
             fi
+
+            success_all=0
             while read -r artf; do
                 echo "download: $(basename $artf)"
-                wget "$tarball_url/$artf" -P "$directory_name/$(dirname $artf)"
+                success=1
+                for dl_src in $dl_source; do
+                    case $dl_src in
+                        $dl_from_stx_mirror)
+                            url="$(url_to_cengn_url "$tarball_url/$artf" "$distro")"
+                            ;;
+                        $dl_from_upstream)
+                            url="$tarball_url/$artf"
+                            ;;
+                        *)
+                            echo "Error: Unknown dl_source '$dl_src'"
+                            continue
+                            ;;
+                    esac
+
+                    wget "$url" -P "$directory_name/$(dirname $artf)"
+                    if [ $? -eq 0 ]; then
+                        success=0
+                        break
+                    else
+                        echo "Warning: Failed to download from $url"
+                        continue;
+                    fi
+                done
+
+                if [ $success -ne 0 ]; then
+                    success_all=1
+                    echo "Error: Failed to download from '$tarball_url/$artf'"
+                fi
             done < "$mvn_artf_file"
+
+            if [ $success_all -ne 0 ]; then
+                echo "Error: Failed to download one or more components from '$tarball_url'"
+                echo "$tarball_url" > "$output_log"
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
 
             # Create tarball
             tar -zcvf "$tarball_name" -C "$directory_name"/ .
@@ -142,7 +294,14 @@ for line in $(cat $tarball_file); do
         elif [[ "$tarball_name" =~ ^'MLNX_OFED_LINUX' ]]; then
             pkg_version=$(echo "$tarball_name" | cut -d "-" -f2-3)
             srpm_path="MLNX_OFED_SRC-${pkg_version}/SRPMS/"
+
             download_package "$tarball_url"
+            if [ $? -ne 0 ]; then
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
             tar -xf "$tarball_name"
             tar -xf "$directory_name/src/MLNX_OFED_SRC-${pkg_version}.tgz"
             # This section of code gets specific SRPMs versions according
@@ -168,8 +327,20 @@ for line in $(cat $tarball_file); do
             rm -rf "$directory_name"
         elif [ "$tarball_name" = "qat1.7.upstream.l.1.0.3-42.tar.gz" ]; then
             download_package $tarball_url
+            if [ $? -ne 0 ]; then
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
         elif [ "$tarball_name" = "tpm-kmod-e6aef069.tar.gz" ]; then
             download_package $tarball_url
+            if [ $? -ne 0 ]; then
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
             tar xf e6aef069b6e97790cb127d5eeb86ae9ff0b7b0e3.tar.gz
             mv linux-tpmdd-e6aef06/drivers/char/tpm $directory_name
             tar czvf $tarball_name $directory_name
@@ -177,22 +348,74 @@ for line in $(cat $tarball_file); do
             rm -rf $directory_name
             rm e6aef069b6e97790cb127d5eeb86ae9ff0b7b0e3.tar.gz
         elif [ "$tarball_name" = "tss2-930.tar.gz" ]; then
-            git clone https://git.code.sf.net/p/ibmtpm20tss/tss ibmtpm20tss-tss
-            pushd ibmtpm20tss-tss
-            git checkout v930
+            dest_dir=ibmtpm20tss-tss
+            for dl_src in $dl_source; do
+                case $dl_src in
+                    $dl_from_stx_mirror)
+                        url="$(url_to_cengn_url "$tarball_url" "$distro")"
+                        ;;
+                    $dl_from_upstream)
+                        url="$tarball_url"
+                        ;;
+                    *)
+                        echo "Error: Unknown dl_source '$dl_src'"
+                        continue
+                        ;;
+                esac
+
+                git clone $url $dest_dir
+                if [ $? -eq 0 ]; then
+                    # Success
+                    break
+                else
+                    echo "Warning: Failed to git clone from '$url'"
+                    continue
+                fi
+            done
+
+            if [ ! -d $dest_dir ]; then
+                echo "Error: Failed to git clone from '$tarball_url'"
+                echo "$tarball_url" > "$output_log"
+                error_count=$((error_count + 1))
+                popd    # pushd $output_tarball
+                continue
+            fi
+
+            pushd $dest_dir
+            branch=$util
+            git checkout $branch
             rm -rf .git
             popd
             mv ibmtpm20tss-tss $directory_name
             tar czvf $tarball_name $directory_name
             rm -rf $directory_name
+            popd     # pushd $dest_dir
         fi
-        popd
+        popd  # pushd $output_tarball
         continue
     fi
 
-    download_cmd="wget -t 5 --wait=15 $tarball_url -O $download_path"
+    if [ -e $download_path ]; then
+        echo "Already have $download_path"
+        continue
+    fi
 
-    if [ ! -e $download_path ]; then
+    for dl_src in $dl_source; do
+        case $dl_src in
+            $dl_from_stx_mirror)
+                url="$(url_to_cengn_url "$tarball_url" "$distro")"
+                ;;
+            $dl_from_upstream)
+                url="$tarball_url"
+                ;;
+            *)
+                echo "Error: Unknown dl_source '$dl_src'"
+                continue
+                ;;
+        esac 
+
+        download_cmd="wget -t 5 --wait=15 $url -O $download_path"
+
         if $download_cmd ; then
             echo "Ok: $download_path"
             pushd $download_directory
@@ -204,16 +427,26 @@ for line in $(cat $tarball_file); do
                 rm -r $directory_name
             fi
             popd
+            break
         else
-            echo "Error: Failed to download $tarball_url" 1>&2
-            echo "$tarball_url" > "$output_log"
+            echo "Warning: Failed to download $url" 1>&2
+            continue
         fi
+    done
 
-    else
-        echo "Already have $download_path"
+    if [ ! -e $download_path ]; then
+        echo "Error: Failed to download $tarball_url" 1>&2
+        echo "$tarball_url" > "$output_log"
+        error_count=$((error_count + 1))
     fi
-
 done
 
 # End of file
 
+if [ $error_count -ne 0 ]; then
+    echo ""
+    echo "Encountered $error_count errors"
+    exit 1
+fi
+
+exit 0
